requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
packaging==25.0
pillow==11.2.1
PyJWT==2.10.1
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
mysqlclient